+++ service/uploadService/form.go
@@ -2,8 +2,8 @@ package uploadService
 
 // FileForm is used when creating or updating a file.
 type FileForm struct {
-	Id     int64  `form:"id"`
-	UserId int64  `form:"userId"`
+	Id     uint   `form:"id"`
+	UserId uint   `form:"userId"`
 	Name   string `form:"name" binding:"required"`
 	Size   int    `form:"size" binding:"required"`
 }
